fix: coerce feature matrix to float64 and add retrain traceback logging
Defends against pandas 3.x / sklearn 1.6 dtype mismatch where dt.hour
returns int32 (not int64) and older GBR validation rejects it. Adds full
traceback to retrain error log for easier future diagnosis.

+++ apps/energy_forecast/model.py
@@ -354,7 +354,7 @@ class EnergyForecastModel:
         else:
             self._feature_medians_by_how = {}
 
-        X = df[feature_cols]            # keep as DataFrame for LightGBM feature names
+        X = df[feature_cols].astype(float)  # coerce int32/extension types → float64 (sklearn/lgb compat)
         y = df["gross_kwh"].to_numpy(dtype=float)
         y_fit = np.log1p(y)             # log-transform reduces influence of rare high peaks
